+++ packages/client-auth/src/sdk/clients.ts
@@ -131,9 +131,8 @@ export class Clients extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }