+++ packages/client-auth/src/sdk/users.ts
@@ -128,9 +128,8 @@ export class Users extends ClientSDK {
             if (nextCursor == null) {
                 return () => null;
             }
-
             const results = dlv(responseData, "cursor.data");
-            if (!results.length) {
+            if (!Array.isArray(results) || !results.length) {
                 return () => null;
             }